==> 03_HyperparameterTuning.qmd <==
---
title: "Script 3 - Hyperparameter Tuning"
author: 
  - name: "MSc. Friederike Johanna Rosa Wölke"
    orcid: "0000-0001-9034-4883"
    url: "https://friedarosa.github.io"
    email: "wolke@fzp.czu.cz"
    corresponding: true
date: "2023-05-29"
---

::: panel-tabset
## Source custom functions

```{r}
#| label: load-packages
#| message: false
#| warning: false
rm(list = ls())
source("src/functions.R")

```

## MachineLearning packages

```{r}
#| label: load-ML-packages
#| message: false
#| error: false

pckgs <- c("dplyr", "ggplot2", "reshape2", 
           "caret", "caretEnsemble", 
           "randomForest", "ranger", "gbm", "xgboost", 
           "gridExtra", "kableExtra")


install_and_load(pckgs)
```

## Load RData/RDS objects to reduce computing time

```{r}
#| label: load-RData
#| message: false
#| error: false

# Load workspace to save computing time:
load("data/RData/03_reduced_hyper_para_tuning_all_models.RData")

# make list of models for evaluation ========
saved_models <- list(Jaccard1 = models_J1, 
                   Jaccard2 = models_J2, 
                   LogRatio1 = models_LR1, 
                   LogRatio2 = models_LR2)

# saveRDS(saved_models, "data/models/reducedModels/03_List_all_reduced_models.rds")

# List of variables to keep in each model
reduced_predictors <- readRDS("data/rds/selected_predictors_list.rds")


H1_vars <- c(
    "sd_PC1", "sd_PC2", # Climatic Niche Breadth
    "GlobRangeSize_m2", "IUCN", "Mass", "Habitat", "Habitat.Density",
    "Migration", "Trophic.Level", "Trophic.Niche", "Primary.Lifestyle",
    "FP", # Phylogenetic Distinctness
    "Hand.Wing.Index") # Measure of dispersal ability
H2_vars <- c(
    "AOO", "rel_occ_Ncells", "mean_prob_cooccur", "D_AOO_a", 
    "moran", "x_intercept", "sp_centr_lon", "sp_centr_lat",
    "lengthMinRect", "widthMinRect", "elonMinRect", "bearingMinRect",
    "circ", "bearing", "Southernness", "Westernness",
    "rel_maxDist", "rel_ewDist", "rel_nsDist", "rel_elonRatio",
    "rel_relCirc", "rel_circNorm", "rel_lin", "Dist_centroid_to_COG",
    "maxDist_toBorder_border", "maxDist_toBorder_centr",
    "minDist_toBorder_centr")
H3_vars <- c("GammaSR", "AlphaSR_sp", "BetaSR_sp")
H4_vars <- c(
    "dataset", "mean_area", "Total_area_samp", "Total_Ncells_samp",
    "mean_cell_length", "atlas_lengthMinRect", "atlas_widthMinRect",
    "atlas_elonMinRect", "atlas_circ", "atlas_bearingMinRect",
    "atlas_bearing", "AtlasCOG_long", "AtlasCOG_lat")

```
:::

### Individual models

#### Hyperparameter tuning

```{r}
#| label: hyperparameter-tuning
#| eval: false
#| code-fold: true

index_list <- list(indices_J1, indices_J2, indices_LR1, indices_LR2)
dat_train_list <- list(dat_train_J1, dat_train_J2, dat_train_LR1, dat_train_LR2)

saved_models <- replicate(4, list())
names(saved_models) <- c("J1", "J2", "LR1", "LR2")

response_list <- c("Jaccard", "Jaccard", "log_R2_1", "log_R2_1")

for(j in c(1:4)){
  ## Loop through differet datasets/Analyses
  indices <- index_list[[j]]
  response <- response_list[[j]] 
  predictors <- rownames(reduced_predictors[[j]])
  
  # Subset the data
  dat_train <- dat_train_list[[j]] %>% 
                    select(any_of(c(response,predictors)))

  
  # Define training control ==========================================================
  trainControl <- trainControl(
    method = "repeatedcv",
    number = 10,
    repeats = 3,
    savePredictions = "final",
    returnResamp = "all",
    verboseIter = FALSE,
    index = indices)

  ## Train ranger model ==========================================================
  set.seed(42)
  tictoc::tic("ranger")
  rangerModel_t <- train(
    as.formula(paste(response, "~ .")),
    data = dat_train,
    method = "ranger",
    trControl = trainControl,
    importance = "permutation",
    scale.permutation.importance = TRUE,
    num.trees = 5000,
    respect.unordered.factors = TRUE,
    oob.error = TRUE,
    tuneLength = 20)
  
  saveRDS(rangerModel_t, paste0("data/models/reducedModels/03_reduced_rangerModel_", response, j, "all.rds"))
  tictoc::toc()
  
  # rangerModel_t <- readRDS( paste0("data/models/reducedModels/rangerModel_", response, j, "all.rds"))

  ### Model results:
  p_rangerModel <- plot(rangerModel_t)
  p_rangerModel
  rangerModel_t$finalModel

  ## Train xgbTree model ==========================================================
  xgb_grid <- expand.grid(
    nrounds = c(1000),
    eta = c(0.1, 0.3),
    max_depth = c(2,3, 5),
    gamma = c(0, 0.01, 0.1),
    colsample_bytree = 0.6,
    min_child_weight = 1,
    subsample =1)
  
  tictoc::tic("xgb")
  set.seed(42)
  xgbModel_t <- train(
    as.formula(paste(response, "~ .")),
    data = dat_train,
    method = "xgbTree",
    trControl = trainControl,
    tuneGrid = xgb_grid)
  saveRDS(xgbModel_t, paste0("data/models/reducedModels/03_reduced_xgbModel_all_", response, j, "TLCUSTOM.rds"))
  tictoc::toc()
  
  # xgbModel_t <- readRDS(paste0("data/models/reducedModels/xgbModel_all_", response, j, "TLCUSTOM.rds"))

  ### Model results:
  p_xgbModel <- plot(xgbModel_t)
  p_xgbModel
  slice_min(xgbModel_t$results, RMSE)
  slice_max(xgbModel_t$results, Rsquared)

  
  ## Train gbm model ==========================================================
  set.seed(42)
  tictoc::tic("gbm")
  gbmModel_t <- train(
    as.formula(paste(response, "~ .")),
    data = dat_train,
    method = "gbm",
    trControl = trainControl,
    tuneLength= 20,
    verbose = FALSE)
  saveRDS(gbmModel_t, paste0("data/models/reducedModels/03_reduced_gbmModel_", response, j, "all.rds"))
  tictoc::toc()

  ### Model results:
  summary.gbm(gbmModel_t$finalModel)
  p_gbmModel <- plot(gbmModel_t)
  p_gbmModel
  gbmModel_t$finalModel
  slice_min(gbmModel_t$results, RMSE)
  slice_max(gbmModel_t$results, Rsquared)
  
  saved_models[[j]] <- list(rangerModel_t, xgbModel_t, gbmModel_t)
}

# saveRDS(saved_models, "data/models/reducedModels/03_List_all_reduced_models.rds")
# save.image("data/RData/03_reduced_hyper_para_tuning.RData")

```

#### Model Evaluation
Here we will extract the model performance, best hyperparameters and variable importances from each model and compare them.

::: panel-tabset
## Jaccard 1

```{r}
# reduced predictor models:

ranger_red <- saved_models[[1]][[1]]
xgb_red <- saved_models[[1]][[2]]
gbm_red <- saved_models[[1]][[3]]

## Ranger =====
ranger_red$finalModel
ranger_red$results
plot(ranger_red)


## XGB ======
plot(xgb_red)
xgb_red


## GBM =====
gbm_red$finalModel
gbm_red

models_J1 <- list(ranger = ranger_red, 
                  xgb = xgb_red,
                  gbm = gbm_red)


## Importance frame (wide format) ===
importances_wide <- cbind(
  varImp(ranger_red)$importance %>% 
    round(2) %>% 
    mutate(var_ranger = rownames(.)) %>% 
    arrange(desc(Overall)) %>% 
    rename("Imp_ranger" = "Overall"),
  
  summary.gbm(gbm_red$finalModel, plotit=FALSE) %>% 
    mutate_if(is.numeric, round, 2) %>% 
    mutate(var_gbm = rownames(.)) %>% 
    arrange(desc(rel.inf)) %>% 
    rename("Imp_gbm" = "rel.inf"),
  
  varImp(xgb_red)$importance %>% 
    round(2) %>% 
    mutate(var_gbm = rownames(.)) %>% 
    arrange(desc(Overall)) %>% 
    rename("Imp_gbm" = "Overall"))  %>% 
  kableExtra::kable()


## Importance frame (long format for plotting) ===
importances_long_J1 <- full_join(
  varImp(ranger_red)$importance %>% 
    round(2) %>% 
    mutate(variable = rownames(.),
           model = "ranger") %>% 
    arrange(desc(Overall)) %>% 
    rename("Importance" = "Overall"),
  summary.gbm(gbm_red$finalModel, plotit=FALSE) %>% 
    mutate_if(is.numeric, round, 2) %>% 
    mutate(variable = rownames(.),
           model = "gbm") %>% 
    arrange(desc(rel.inf)) %>% 
    rename("Importance" = "rel.inf") %>% 
    select(variable, model, Importance)) %>% 
  full_join(varImp(xgb_red)$importance %>% 
    round(2) %>% 
    mutate(variable = rownames(.),
           model = "xgb") %>% 
    arrange(desc(Overall)) %>% 
    rename("Importance" = "Overall")) %>%
  mutate(hypothesis = case_when(variable %in% H1_vars ~ "H1",
                                variable %in% H2_vars ~ "H2",
                                variable %in% H3_vars ~ "H3",
                                variable %in% H4_vars ~ "H4",
                                .default = "H1"))
```

## Jaccard 2

```{r}
#| eval: true
# Jaccard 2
ranger_red <- saved_models[[2]][[1]]
xgb_red <- saved_models[[2]][[2]]
gbm_red <- saved_models[[2]][[3]]

## Ranger =====
ranger_red$finalModel
ranger_red$results
plot(ranger_red)


## XGB ======
plot(xgb_red)
xgb_red


## GBM =====
gbm_red$finalModel
gbm_red

models_J2 <- list(ranger = ranger_red, 
                  xgb = xgb_red,
                  gbm = gbm_red)






## Importance frame (wide format) ===
importances_wide <- cbind(
  varImp(ranger_red)$importance %>% 
    round(2) %>% 
    mutate(var_ranger = rownames(.)) %>% 
    arrange(desc(Overall)) %>% 
    rename("Imp_ranger" = "Overall"),
  
  summary.gbm(gbm_red$finalModel, plotit=FALSE) %>% 
    mutate_if(is.numeric, round, 2) %>% 
    mutate(var_gbm = rownames(.)) %>% 
    arrange(desc(rel.inf)) %>% 
    rename("Imp_gbm" = "rel.inf"),
  
  varImp(xgb_red)$importance %>% 
    round(2) %>% 
    mutate(var_gbm = rownames(.)) %>% 
    arrange(desc(Overall)) %>% 
    rename("Imp_gbm" = "Overall"))  %>% 
  kableExtra::kable()


## Importance frame (long format for plotting) ===
importances_long_J2 <- full_join(
  varImp(ranger_red)$importance %>% 
    round(2) %>% 
    mutate(variable = rownames(.),
           model = "ranger") %>% 
    arrange(desc(Overall)) %>% 
    rename("Importance" = "Overall"),
  
  summary.gbm(gbm_red$finalModel, plotit=FALSE) %>% 
    mutate_if(is.numeric, round, 2) %>% 
    mutate(variable = rownames(.),
           model = "gbm") %>% 
    arrange(desc(rel.inf)) %>% 
    rename("Importance" = "rel.inf") %>% 
    select(variable, model, Importance)) %>% 
  
  full_join(varImp(xgb_red)$importance %>% 
    round(2) %>% 
    mutate(variable = rownames(.),
           model = "xgb") %>% 
    arrange(desc(Overall)) %>% 
    rename("Importance" = "Overall")) %>%
  mutate(hypothesis = case_when(variable %in% H1_vars ~ "H1",
                                variable %in% H2_vars ~ "H2",
                                variable %in% H3_vars ~ "H3",
                                variable %in% H4_vars ~ "H4",
                                .default = "H1"))


```

## Log Ratio 1

```{r}
## reduced:
ranger_red <- saved_models[[3]][[1]]
xgb_red <- saved_models[[3]][[2]]
gbm_red <- saved_models[[3]][[3]]

## Ranger =====
ranger_red$finalModel
ranger_red$results
plot(ranger_red)


## XGB ======
plot(xgb_red)
xgb_red


## GBM =====
gbm_red$finalModel
gbm_red

models_LR1 <- list(ranger = ranger_red, 
                  xgb = xgb_red,
                  gbm = gbm_red)






## Importance frame (wide format) ===
importance_wide <- cbind(
  varImp(ranger_red)$importance %>% 
    round(2) %>% 
    mutate(var_ranger = rownames(.)) %>% 
    arrange(desc(Overall)) %>% 
    rename("Imp_ranger" = "Overall"),
  summary.gbm(gbm_red$finalModel, plotit=FALSE) %>% 
    mutate_if(is.numeric, round, 2) %>% 
    mutate(var_gbm = rownames(.)) %>% 
    arrange(desc(rel.inf)) %>% 
    rename("Imp_gbm" = "rel.inf"),
  varImp(xgb_red)$importance %>% 
    round(2) %>% 
    mutate(var_gbm = rownames(.)) %>% 
    arrange(desc(Overall)) %>% 
    rename("Imp_gbm" = "Overall"))  %>% 
  kableExtra::kable()


## Importance frame (long format for plotting) ===
importances_long_LR1 <- full_join(
  varImp(ranger_red)$importance %>% 
    round(2) %>% 
    mutate(variable = rownames(.),
           model = "ranger") %>% 
    arrange(desc(Overall)) %>% 
    rename("Importance" = "Overall"),
  summary.gbm(gbm_red$finalModel, plotit=FALSE) %>% 
    mutate_if(is.numeric, round, 2) %>% 
    mutate(variable = rownames(.),
           model = "gbm") %>% 
    arrange(desc(rel.inf)) %>% 
    rename("Importance" = "rel.inf") %>% 
    select(variable, model, Importance)) %>% 
  full_join(varImp(xgb_red)$importance %>% 
    round(2) %>% 
    mutate(variable = rownames(.),
           model = "xgb") %>% 
    arrange(desc(Overall)) %>% 
    rename("Importance" = "Overall")) %>%
  mutate(hypothesis = case_when(variable %in% H1_vars ~ "H1",
                                variable %in% H2_vars ~ "H2",
                                variable %in% H3_vars ~ "H3",
                                variable %in% H4_vars ~ "H4",
                                .default = "H1"))

```

## Log Ratio 2

```{r}
## reduced:
ranger_red <- saved_models[[4]][[1]]
xgb_red <- saved_models[[4]][[2]]
gbm_red <- saved_models[[4]][[3]]

## Ranger =====
ranger_red$finalModel
ranger_red$results
plot(ranger_red)


## XGB ======
plot(xgb_red)
xgb_red


## GBM =====
gbm_red$finalModel
gbm_red

models_LR2 <- list(ranger = ranger_red, 
                  xgb = xgb_red,
                  gbm = gbm_red)




## Importance frame (wide format) ===
importance_wide <- cbind(
  varImp(ranger_red)$importance %>% 
    round(2) %>% 
    mutate(var_ranger = rownames(.)) %>% 
    arrange(desc(Overall)) %>% 
    rename("Imp_ranger" = "Overall"),
  summary.gbm(gbm_red$finalModel, plotit=FALSE) %>% 
    mutate_if(is.numeric, round, 2) %>% 
    mutate(var_gbm = rownames(.)) %>% 
    arrange(desc(rel.inf)) %>% 
    rename("Imp_gbm" = "rel.inf"),
  varImp(xgb_red)$importance %>% 
    round(2) %>% 
    mutate(var_gbm = rownames(.)) %>% 
    arrange(desc(Overall)) %>% 
    rename("Imp_gbm" = "Overall")
)  %>% kableExtra::kable()


## Importance frame (long format for plotting) ===
importances_long_LR2 <- full_join(
  varImp(ranger_red)$importance %>% 
    round(2) %>% 
    mutate(variable = rownames(.),
           model = "ranger") %>% 
    arrange(desc(Overall)) %>% 
    rename("Importance" = "Overall"),
  summary.gbm(gbm_red$finalModel, plotit=FALSE) %>% 
    mutate_if(is.numeric, round, 2) %>% 
    mutate(variable = rownames(.),
           model = "gbm") %>% 
    arrange(desc(rel.inf)) %>% 
    rename("Importance" = "rel.inf") %>% 
    select(variable, model, Importance)) %>% 
  full_join(varImp(xgb_red)$importance %>% 
    round(2) %>% 
    mutate(variable = rownames(.),
           model = "xgb") %>% 
    arrange(desc(Overall)) %>% 
    rename("Importance" = "Overall")) %>%
  mutate(hypothesis = case_when(variable %in% H1_vars ~ "H1",
                                variable %in% H2_vars ~ "H2",
                                variable %in% H3_vars ~ "H3",
                                variable %in% H4_vars ~ "H4",
                                .default = "H1"))



```
:::

### Variable Importances across models
::: panel-tabset
## Jaccard 1
```{r}

# Plot variable importances across models:
ggplot(data = importances_long_J1, 
       aes(y = variable, x = Importance, fill = hypothesis))+
  geom_col()+
  facet_grid(~model, scales = "free")+
  theme_classic()+
  scale_fill_manual(values = c("seagreen3", "mediumpurple", "darkorange", "gold"))
```
## Jaccard 2
```{r}

# Plot variable importances across models:
ggplot(data = importances_long_J2, 
       aes(y = variable, x = Importance, fill = hypothesis))+
  geom_col()+
  facet_grid(~model, scales = "free")+
  theme_classic()+
  scale_fill_manual(values = c("seagreen3", "mediumpurple", "darkorange", "gold"))
```

## Log Ratio 1
```{r}
# Plot variable importances across models:
ggplot(data = importances_long_LR1, 
       aes(y = variable, x = Importance, fill = hypothesis))+
  geom_col()+
  facet_grid(~model, scales = "free")+
  theme_classic()+
  scale_fill_manual(values = c("seagreen3", "mediumpurple", "darkorange", "gold"))
```

## Log Ratio 2
```{r}
# Plot variable importances across models:
ggplot(data = importances_long_LR2, 
       aes(y = variable, x = Importance, fill = hypothesis))+
  geom_col()+
  facet_grid(~model, scales = "free")+
  theme_classic()+
  scale_fill_manual(values = c("seagreen3", "mediumpurple", "darkorange", "gold"))
```


:::




# Resamples across models:
:::panel-tabset
## Jaccard 1
```{r}
## Jaccard 1
# Create resamples for each model
resamps_J1 <- resamples(list(
  ranger = models_J1[[1]],
  xgb = models_J1[[2]],
  gbm = models_J1[[3]]
))

# Plot the resampled error rates for each model
dotplot_resamps_J1 <- dotplot(resamps_J1)

# Summarize the resamples
summary_resamps_J1 <- summary(resamps_J1)
```

## Jaccard 2
```{r}
# Create resamples for each model
resamps_J2 <- resamples(list(
  ranger = models_J2[[1]],
  xgb = models_J2[[2]],
  gbm = models_J2[[3]]
))

# Plot the resampled error rates for each model
dotplot_resamps_J2 <- dotplot(resamps_J2)

# Summarize the resamples
summary_resamps_J2 <- summary(resamps_J2)
```

## Log Ratio 1
```{r}
# Create resamples for each model
resamps_LR1 <- resamples(list(
  ranger = models_LR1[[1]],
  xgb = models_LR1[[2]],
  gbm = models_LR1[[3]]
))

# Plot the resampled error rates for each model
dotplot_resamps_LR1 <- dotplot(resamps_LR1)

# Summarize the resamples
summary_resamps_LR1 <- summary(resamps_LR1)
```

## Log Ratio 2
```{r}
# Create resamples for each model
resamps_LR2 <- resamples(list(
  ranger = models_LR2[[1]],
  xgb = models_LR2[[2]],
  gbm = models_LR2[[3]]
))

# Plot the resampled error rates for each model
dotplot_resamps_LR2 <- dotplot(resamps_LR2)

# Summarize the resamples
summary_resamps_LR2 <- summary(resamps_LR2)
```
:::

### Summary across all analyses
```{r}
#| code-fold: true
resamples_all <- list(
  Jaccard1 = list(
    resamps_J1 = resamps_J1,
    Dotplot = dotplot_resamps_J1, # store the dotplot object
    Summary = summary_resamps_J1  # store the summary object
  ),
  Jaccard2 = list(
    resamps_J2 = resamps_J2,
    Dotplot = dotplot_resamps_J2, # store the dotplot object
    Summary = summary_resamps_J2  # store the summary object
  ),
  LogRatio1 = list(
    resamps_LR1 = resamps_LR1,
    Dotplot = dotplot_resamps_LR1, # store the dotplot object
    Summary = summary_resamps_LR1  # store the summary object
  ),
  LogRatio2 = list(
    resamps_LR2 = resamps_LR2,
    Dotplot = dotplot_resamps_LR2, # store the dotplot object
    Summary = summary_resamps_LR2  # store the summary object
))


resamples_all


# save.image("data/RData/03_reduced_hyper_para_tuning_all_models.RData")
```
